Reorganize and add tests
