+++ setup.cfg
@@ -142,6 +142,7 @@ ignore =
     S603
     W503 # Line break before binary operator (flake8 is wrong)
     E203  # whitespace before ':'
+    S113 # Requests call without timeout
 exclude =
     .tox,
     .git,
@@ -155,6 +156,8 @@ exclude =
     .cache,
     .eggs,
     data
+    src/curies/mapping_service/rdflib_custom.py
+    src/curies/mapping_service/rdflib_patches.py
 per-file-ignores =
     src/curies/cli.py:DAR101,DAR201
 max-line-length = 120
